Prevent overwriting of figure file

+++ misc/latex_table_creator.py
@@ -198,8 +198,9 @@ def make_picture(tables):
             # legend_str = ",".join(figure["Axis"].keys())
             # figure_str += f"    \\legend{{{legend_str}}}\n"
             figure_str += "    \\end{axis}\n\\end{tikzpicture}"
-            with open(path_figure, "w") as fp:
-                fp.write(figure_str)
+            if not path_figure.exists():
+                with open(path_figure, "w") as fp:
+                    fp.write(figure_str)
 
 
 def main():
